# api/openapi.yaml
openapi: 3.1.0
info:
  title: Flex API
  version: 0.5.0
  description: API du serveur multimédia personnel Flex.
servers:
  - url: /api
security:
  - cookieAuth: []
paths:
...
  /auth/setup:
    post:
      security: []
      operationId: setupAdministrator
      summary: Crée le premier administrateur et ouvre une session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupAuthRequest"
      responses:
        "201":
          description: Administrateur créé.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Flex est déjà configuré.
        "422":
          description: Informations invalides.
  /auth/login:
    post:
      security: []
      operationId: login
      summary: Ouvre une session utilisateur.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Session ouverte.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Identifiants incorrects.
...
  /auth/password:
    put:
      operationId: changePassword
      summary: Modifie le mot de passe de l'utilisateur connecté et révoque ses autres sessions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Mot de passe modifié et session courante renouvelée.
        "401":
          description: Mot de passe actuel incorrect.
        "422":
          description: Nouveau mot de passe invalide.
  /auth/profile:
    patch:
      operationId: updateProfile
      summary: Modifie le profil de l'utilisateur connecté.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profil modifié.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Nom d’utilisateur déjà utilisé.
        "422":
          description: Nom d’utilisateur invalide.
  /users:
    get:
      operationId: listUsers
      summary: Liste les comptes Flex. Réservé aux administrateurs.
      responses:
        "200":
          description: Liste des utilisateurs.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuthUser"
        "403":
          description: Droits administrateur requis.
    post:
      operationId: createUser
      summary: Crée un compte Flex. Réservé aux administrateurs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Utilisateur créé.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Nom d’utilisateur déjà utilisé.
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateUser
      summary: Modifie un compte Flex. Réservé aux administrateurs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Utilisateur modifié.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Le dernier administrateur actif doit être conservé.
    delete:
      operationId: deleteUser
      summary: Supprime un compte et ses sessions. Réservé aux administrateurs.
      responses:
        "204":
          description: Utilisateur supprimé.
        "409":
          description: Le dernier administrateur actif doit être conservé.
  /users/{userId}/password:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
    put:
      operationId: resetUserPassword
      summary: Remplace le mot de passe et révoque les sessions du compte.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [password]
              properties:
                password:
                  type: string
                  minLength: 12
                  maxLength: 256
      responses:
        "204":
          description: Mot de passe remplacé et sessions révoquées.
...
  /libraries:
    get:
      operationId: listLibraries
      summary: Liste les bibliothèques configurées.
      responses:
        "200":
          description: Liste des bibliothèques.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Library"
    post:
      operationId: createLibrary
      summary: Ajoute un dossier de médias à Flex.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLibraryRequest"
      responses:
        "201":
          description: Bibliothèque créée.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Library"
        "409":
          description: Ce chemin est déjà configuré.
        "422":
          description: Le nom ou le chemin est invalide.
  /libraries/{libraryId}/scan:
    parameters:
      - $ref: "#/components/parameters/LibraryId"
    post:
      operationId: scanLibrary
      summary: Analyse les vidéos présentes dans une bibliothèque.
      responses:
        "200":
          description: Résultat de l'analyse.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
        "404":
          description: Bibliothèque introuvable.
    get:
      operationId: getLibraryScanStatus
      summary: Retourne l'état de l'analyse d'une bibliothèque.
      responses:
        "200":
          description: État actuel de l'analyse.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanStatus"
  /libraries/{libraryId}/folders:
    parameters:
      - $ref: "#/components/parameters/LibraryId"
    get:
      operationId: listLibraryFolders
      summary: Liste les dossiers relatifs des vidéos d'une bibliothèque.
      responses:
        "200":
          description: Association entre chaque vidéo et son dossier relatif.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/FolderAssignment"
  /libraries/{libraryId}:
    parameters:
      - $ref: "#/components/parameters/LibraryId"
    patch:
      operationId: updateLibrary
      summary: Modifie le nom ou le dossier d'une bibliothèque.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLibraryRequest"
      responses:
        "200":
          description: Bibliothèque mise à jour.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Library"
        "404":
          description: Bibliothèque introuvable.
    delete:
      operationId: deleteLibrary
      summary: Retire une bibliothèque et son index sans supprimer ses fichiers.
      responses:
        "204":
          description: Bibliothèque retirée.
        "404":
          description: Bibliothèque introuvable.
  /media:
    get:
      operationId: listMedia
      summary: Liste les vidéos indexées d'une bibliothèque.
      parameters:
        - $ref: "#/components/parameters/LibraryIdQuery"
      responses:
        "200":
          description: Liste des vidéos.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/MediaFile"
  /collections:
    get:
      operationId: listCollections
      summary: Liste les collections personnelles.
      responses:
        "200":
          description: Liste des collections.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionList"
    post:
      operationId: createCollection
      summary: Crée une collection.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollectionRequest"
      responses:
        "201":
          description: Collection créée.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "409":
          description: Une collection avec ce nom existe déjà.
        "422":
          description: Nom invalide.
  /collections/{collectionId}:
    parameters:
      - name: collectionId
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateCollection
      summary: Renomme une collection.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollectionRequest"
      responses:
        "200":
          description: Collection modifiée.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "404":
          description: Collection introuvable.
        "409":
          description: Une collection avec ce nom existe déjà.
        "422":
          description: Nom invalide.
    delete:
      operationId: deleteCollection
      summary: Supprime une collection sans supprimer ses vidéos.
      responses:
        "204":
          description: Collection supprimée.
        "404":
          description: Collection introuvable.
  /collections/{collectionId}/media:
    parameters:
      - name: collectionId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: listCollectionMedia
      summary: Liste les identifiants des vidéos d'une collection.
      responses:
        "200":
          description: Identifiants des vidéos.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [mediaIds]
                properties:
                  mediaIds:
                    type: array
                    items:
                      type: string
  /collections/{collectionId}/media/{mediaId}:
    parameters:
      - name: collectionId
        in: path
        required: true
        schema:
          type: string
      - $ref: "#/components/parameters/MediaId"
    delete:
      operationId: removeMediaFromCollection
      summary: Retire une vidéo d'une collection sans supprimer la vidéo.
      responses:
        "204":
          description: Vidéo retirée de la collection.
        "404":
          description: Association introuvable.
  /media/{mediaId}/collections:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      operationId: listMediaCollections
      summary: Liste les collections contenant une vidéo.
      responses:
        "200":
          description: Collections de la vidéo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionList"
    put:
      operationId: setMediaCollections
      summary: Remplace les collections associées à une vidéo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetMediaCollectionsRequest"
      responses:
        "200":
          description: Associations enregistrées.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionList"
        "404":
          description: Vidéo introuvable.
  /media/{mediaId}:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      operationId: getMedia
      summary: Retourne les informations d'une vidéo.
      responses:
        "200":
          description: Informations de la vidéo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaFile"
        "404":
          description: Vidéo introuvable.
    patch:
      operationId: updateMedia
      summary: Modifie les métadonnées personnelles d'une vidéo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMediaRequest"
      responses:
        "200":
          description: Métadonnées enregistrées.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaFile"
        "404":
          description: Vidéo introuvable.
        "422":
          description: Métadonnées invalides.
  /media/{mediaId}/favorite:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    put:
      operationId: setMediaFavorite
      summary: Ajoute ou retire une vidéo des favoris de l'utilisateur connecté.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetFavoriteRequest"
      responses:
        "200":
          description: Favori enregistré pour l'utilisateur connecté.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaFile"
        "404":
          description: Vidéo introuvable.
...
  /search:
    get:
      operationId: searchMedia
      summary: Recherche des vidéos par titre ou nom de fichier.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Résultats de la recherche.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/MediaSearchResult"
  /tags:
    get:
      operationId: listTags
      summary: Liste les tags disponibles.
      responses:
        "200":
          description: Liste des tags.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
    post:
      operationId: createTag
      summary: Crée un tag personnalisé.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTagRequest"
      responses:
        "201":
          description: Tag créé.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "409":
          description: Un tag avec ce nom existe déjà.
        "422":
          description: Nom ou couleur invalide.
...
  /media/{mediaId}/tags:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      operationId: listMediaTags
      summary: Liste les tags attribués à une vidéo.
      responses:
        "200":
          description: Tags de la vidéo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
        "404":
          description: Vidéo introuvable.
    put:
      operationId: setMediaTags
      summary: Remplace les tags attribués à une vidéo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetMediaTagsRequest"
      responses:
        "200":
          description: Tags enregistrés.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
        "404":
          description: Vidéo introuvable.
        "422":
          description: Sélection invalide.
  /media/{mediaId}/thumbnail:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      operationId: getMediaThumbnail
      summary: Génère ou retourne la miniature mise en cache d'une vidéo.
      responses:
        "200":
          description: Miniature JPEG.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
  /media/{mediaId}/stream:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      operationId: streamMedia
      summary: Lit directement le fichier vidéo avec prise en charge HTTP Range.
      responses:
        "200":
          description: Fichier vidéo complet.
        "206":
          description: Portion du fichier demandée par le lecteur.
  /media/{mediaId}/playback:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      operationId: getMediaPlayback
      summary: Sélectionne la stratégie de lecture de la vidéo.
      responses:
        "200":
          description: Lecture directe ou manifeste HLS transcodé.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackInfo"
  /media/{mediaId}/hls/{filename}:
    parameters:
      - $ref: "#/components/parameters/MediaId"
      - name: filename
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getMediaHLSFile
      summary: Génère puis retourne un manifeste ou segment HLS.
      responses:
        "200":
          description: Manifeste M3U8 ou segment MPEG-TS.
  /media/{mediaId}/progress:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      operationId: getPlaybackProgress
      summary: Retourne la progression de lecture locale.
      responses:
        "200":
          description: Progression actuelle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackProgress"
    put:
      operationId: savePlaybackProgress
      summary: Sauvegarde la position de lecture locale.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavePlaybackProgress"
      responses:
        "200":
          description: Progression sauvegardée.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackProgress"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: flex_session
  parameters:
    LibraryId:
      name: libraryId
      in: path
      required: true
      schema:
        type: string
    LibraryIdQuery:
      name: libraryId
      in: query
      required: true
      schema:
        type: string
...
  schemas:
    AuthUser:
      type: object
      additionalProperties: false
      required: [id, username, role, active]
      properties:
        id:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        active:
          type: boolean
...
    SetupAuthRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 12
          maxLength: 256
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 12
          maxLength: 256
    UpdateProfileRequest:
      type: object
      additionalProperties: false
      required: [username]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [username, password, role]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 12
          maxLength: 256
        role:
          type: string
          enum: [admin, user]
    UpdateUserRequest:
      type: object
      additionalProperties: false
      required: [username, role, active]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        role:
          type: string
          enum: [admin, user]
        active:
          type: boolean
...
    Library:
      type: object
      additionalProperties: false
      required: [id, name, path, createdAt, lastScanAt, lastScanDiscovered, lastScanIndexed, lastScanUnchanged, lastScanSkipped]
      properties:
        id:
          type: string
        name:
          type: string
        path:
          type: string
        createdAt:
          type: string
          format: date-time
        lastScanAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        lastScanDiscovered:
          type: integer
          minimum: 0
        lastScanIndexed:
          type: integer
          minimum: 0
        lastScanUnchanged:
          type: integer
          minimum: 0
        lastScanSkipped:
          type: integer
          minimum: 0
    CreateLibraryRequest:
      type: object
      additionalProperties: false
      required: [name, path]
      properties:
        name:
          type: string
          minLength: 1
        path:
          type: string
          minLength: 1
    ScanResult:
      type: object
      additionalProperties: false
      required: [discovered, indexed, unchanged, removed, skipped, issues]
      properties:
        discovered:
          type: integer
          minimum: 0
        indexed:
          type: integer
          minimum: 0
        unchanged:
          type: integer
          minimum: 0
        removed:
          type: integer
          minimum: 0
        skipped:
          type: integer
          minimum: 0
        issues:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/ScanIssue"
...
    ScanStatus:
      type: object
      additionalProperties: false
      required: [state]
      properties:
        state:
          type: string
          enum: [idle, pending, scanning, completed, failed]
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        result:
          $ref: "#/components/schemas/ScanResult"
        lastError:
          type: string
    FolderAssignment:
      type: object
      additionalProperties: false
      required: [mediaId, folder]
      properties:
        mediaId:
          type: string
        folder:
          type: string
...
    Collection:
      type: object
      additionalProperties: false
      required: [id, name, mediaCount]
      properties:
        id:
          type: string
        name:
          type: string
        mediaCount:
          type: integer
          minimum: 0
    CollectionList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Collection"
    CreateCollectionRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
...
    MediaFile:
      type: object
      required: [id, libraryId, filename, sizeBytes, durationMs, width, height, container, videoCodec, audioCodec, modifiedAt, progressMs, completed, title, description, recordedAt, favorite]
      properties:
        id:
          type: string
        libraryId:
          type: string
        filename:
          type: string
        sizeBytes:
          type: integer
          format: int64
        durationMs:
          type: integer
          format: int64
        width:
          type: integer
        height:
          type: integer
        container:
          type: string
        videoCodec:
          type: string
        audioCodec:
          type: string
        modifiedAt:
          type: string
          format: date-time
        progressMs:
          type: integer
          format: int64
          minimum: 0
        completed:
          type: boolean
        title:
          type: string
        description:
          type: string
        recordedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        favorite:
          type: boolean
...